An example about how to enable policy enforcement to Elytron/Jakarta applications
Closes #keycloak#19540
