+++ app-authz-jakarta-servlet/src/main/java/org/keycloak/quickstart/Controller.java
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2016 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.keycloak.quickstart;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+import java.io.IOException;
+
+/**
+ * Controller simplifies access to the server environment from the JSP.
+ *
+ * @author Stan Silvert ssilvert@redhat.com (C) 2015 Red Hat Inc.
+ */
+public class Controller {
+
+    public boolean isLoggedIn(HttpServletRequest req) {
+        return req.getUserPrincipal() != null;
+    }
+
+    public void handleLogout(HttpServletRequest req, HttpServletResponse res) throws ServletException, IOException {
+        if (isLogoutAction(req)) {
+            req.logout();
+            res.sendRedirect(req.getContextPath());
+        }
+    }
+
+    public boolean isLogoutAction(HttpServletRequest req) {
+        return getAction(req).equals("logout");
+    }
+
+    private String getAction(HttpServletRequest req) {
+        if (req.getParameter("action") == null) return "";
+        return req.getParameter("action");
+    }
+}
