+++ app-authz-jakarta-servlet/src/main/webapp/WEB-INF/oidc.json
@@ -1,0 +1,9 @@
+{
+  "realm": "quickstart-authz-servlet",
+  "auth-server-url": "http://127.0.0.1:8180/auth",
+  "ssl-required": "external",
+  "resource": "authz-servlet",
+  "credentials": {
+    "secret": "secret"
+  }
+}
