+++ app-authz-jakarta-servlet/src/main/webapp/WEB-INF/policy-enforcer.json
@@ -1,0 +1,7 @@
+{
+  "path-cache": {
+    "lifespan": 0,
+    "max-entries": 0
+  },
+  "on-deny-redirect-to" : "/authz-servlet/accessDenied.jsp"
+}
