+++ app-authz-jakarta-servlet/src/test/resources/oidc.json
@@ -1,0 +1,9 @@
+{
+  "realm": "quickstart-authz-servlet",
+  "auth-server-url": "http://localhost:8180/auth",
+  "ssl-required": "external",
+  "resource": "authz-servlet",
+  "credentials": {
+    "secret": "secret"
+  }
+}
