--- app-authz-jee-servlet/config/keycloak-example.json
@@ -1,16 +1,0 @@
-{
-  "realm": "quickstart",
-  "auth-server-url": "http://localhost:8180/auth",
-  "ssl-required": "external",
-  "resource": "authz-servlet",
-  "credentials": {
-    "secret": "secret"
-  },
-  "policy-enforcer": {
-    "path-cache": {
-      "lifespan": 0,
-      "max-entries": 0
-    },
-    "on-deny-redirect-to" : "/authz-servlet/accessDenied.jsp"
-  }
-}
