--- app-authz-jee-servlet/src/test/resources/arquillian.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- JBoss, Home of Professional Open Source Copyright 2012, Red Hat, Inc.
-	and/or its affiliates, and individual contributors by the @authors tag. See
-	the copyright.txt in the distribution for a full listing of individual contributors.
-	Licensed under the Apache License, Version 2.0 (the "License"); you may not
-	use this file except in compliance with the License. You may obtain a copy
-	of the License at http://www.apache.org/licenses/LICENSE-2.0 Unless required
-	by applicable law or agreed to in writing, software distributed under the
-	License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS
-	OF ANY KIND, either express or implied. See the License for the specific
-	language governing permissions and limitations under the License. -->
-<arquillian xmlns="http://jboss.org/schema/arquillian"
-            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-            xsi:schemaLocation="http://jboss.org/schema/arquillian
-        http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
-
-
-    <!-- Force the use of the Servlet 3.0 protocol with all containers, as it
-        is the most mature -->
-    <defaultProtocol type="Servlet 3.0" />
-
-    <!-- Example configuration for a remote JBoss AS 7 instance -->
-    <container qualifier="jboss" default="true">
-        <!-- If you want to use the JBOSS_HOME environment variable, just delete
-            the jbossHome property -->
-        <configuration>
-            <property name="managementAddress">127.0.0.1</property>
-            <property name="managementPort">9990</property>
-            <property name="jbossHome">${basedir}/target/wildfly-${version.wildfly}</property>
-        </configuration>
-    </container>
-
-    <extension qualifier="webdriver">
-        <property name="browser">phantomjs</property>
-    </extension>
-</arquillian>
