--- app-authz-jee-servlet/src/test/resources/keycloak.json
@@ -1,12 +1,0 @@
-{
-  "realm": "quickstart-authz-servlet",
-  "auth-server-url": "http://localhost:8180/auth",
-  "ssl-required": "external",
-  "resource": "authz-servlet",
-  "credentials": {
-    "secret": "secret"
-  },
-  "policy-enforcer": {
-    "on-deny-redirect-to" : "/authz-servlet/accessDenied.jsp"
-  }
-}
